Messages are now sent to the api and stored to the database * Some other small fixes

+++ apps/message-crawler/telegram-crawler-core/src/main/java/io/d2a/schwurbelwatch/tgcrawler/api/messages/Message.java
@@ -6,8 +6,8 @@ import lombok.ToString;
 @ToString
 public class Message {
 
-  public int messageId;
-  public int chatId;
+  public long messageId;
+  public long chatId;
   public int userId;
 
   @SerializedName("content_type")
